Refactored to sit inside 'eLCS' module (classes are now submodules)
